requirements.txt
Flask
inference-sdk
Pillow